// src/generated/TMS_v2/schema/transport-action-type.ts
/*
 * Copyright (c) 2025 SAP SE or an SAP affiliate company. All rights reserved.
 *
 * This is a generated file powered by the SAP Cloud SDK for JavaScript.
 */

/**
 * Action type:
 *   * `FU` - File upload
 *   * `NE` - Node export
 *   * `NU` - Node upload
 *   * `I` - Import
 *   * `DQ` - Delete queue entry
 *   * `RQ` - Reset queue entry
 *   * `CQ` - Cleanup queue entry
 *   * `FQ` - Forward queue entry
 *   * `AQ` - Add queue entry
 *
 */
export type TransportActionType =
  | 'FU'
  | 'NE'
  | 'NU'
  | 'I'
  | 'DQ'
  | 'RQ'
  | 'CQ'
  | 'FQ'
  | 'AQ';