// src/generated/TMS_v2/schema/action-status.ts
/*
 * Copyright (c) 2025 SAP SE or an SAP affiliate company. All rights reserved.
 *
 * This is a generated file powered by the SAP Cloud SDK for JavaScript.
 */

/**
 * Action status
 *   * `succeeded`
 *   * `warning`
 *   * `error`
 *   * `fatal`
 *   * `running`
 *   * `initial`
 *   * `unknown`
 *
 */
export type ActionStatus =
  | 'succeeded'
  | 'warning'
  | 'error'
  | 'fatal'
  | 'running'
  | 'initial'
  | 'unknown';